refactor : repackaging domian
요금관련 객체 공통으로 이동
