+++ src/main/java/nextstep/subway/common/domain/DiscountByAge.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.common.domain;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
