+++ src/main/java/nextstep/subway/line/domain/Surcharge.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.dto;
+package nextstep.subway.line.domain;
 
 import javax.persistence.Embeddable;
 import java.math.BigDecimal;
