+++ src/test/java/nextstep/subway/line/domain/SurchargeByLineTest.java
@@ -1,5 +1,6 @@
 package nextstep.subway.line.domain;
 
+import nextstep.subway.common.domain.SurchargeByLine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
