+++ src/test/java/nextstep/subway/path/PathAcceptanceTest.java
@@ -5,12 +5,12 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.auth.dto.TokenRequest;
+import nextstep.subway.common.domain.DiscountByAge;
 import nextstep.subway.common.domain.SubwayFare;
+import nextstep.subway.common.domain.SurchargeByLine;
 import nextstep.subway.line.acceptance.LineAcceptanceTest;
-import nextstep.subway.line.domain.SurchargeByLine;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.path.domain.DiscountByAge;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
