+++ src/test/java/nextstep/subway/path/domain/DiscountByAgeTest.java
@@ -1,5 +1,6 @@
 package nextstep.subway.path.domain;
 
+import nextstep.subway.common.domain.DiscountByAge;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
